+++ src/vite-env.d.ts
@@ -1,1 +1,4 @@
 /// <reference types="vite/client" />
+declare const __APP_VERSION__: string;
+declare const __APP_NAME__: string;
+declare const __APP_DESCRIPTION__: string;
