+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/layout/AppLayout.tsx","./src/components/layout/ChannelList.tsx","./src/components/layout/ChatArea.tsx","./src/components/layout/MemberList.tsx","./src/components/layout/PrivateMessagesList.tsx","./src/components/layout/ServerList.tsx","./src/components/ui/AddServerModal.tsx","./src/components/ui/EmojiSelector.tsx","./src/components/ui/UserSettings.tsx","./src/lib/ircClient.ts","./src/lib/ircUtils.ts","./src/protocol/index.ts","./src/protocol/isupport.ts","./src/store/index.ts","./src/types/index.ts","./vite.config.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/layout/applayout.tsx","./src/components/layout/channellist.tsx","./src/components/layout/chatarea.tsx","./src/components/layout/memberlist.tsx","./src/components/layout/serverlist.tsx","./src/components/ui/addservermodal.tsx","./src/components/ui/emojiselector.tsx","./src/components/ui/usersettings.tsx","./src/lib/ircclient.ts","./src/lib/ircutils.ts","./src/protocol/index.ts","./src/protocol/isupport.ts","./src/store/index.ts","./src/types/index.ts","./vite.config.ts"],"errors":true,"version":"5.6.3"}
